fixed shifts, tests before optomised the op away.  Added implicit int (and others) cast

+++ Compiler/src/cs444/parser/symbols/ast/INumericLiteral.java
@@ -3,5 +3,5 @@ package cs444.parser.symbols.ast;
 import cs444.parser.symbols.ISymbol;
 
 public interface INumericLiteral extends ISymbol{
-    public int getValue();
+    public long getValue();
 }
